(HC-41) Port Mergeable* classes and relevant methods
This commit ports pieces of the ConfigDelayedMergeObject and
ConfigDelayedMerge classes which are necessary to implement
many Merge* methods elsewhere.

Several of these methods are also ported as part of this commit.
They include:
ConfigValue::mergedWithTheUnmergeable,
ConfigValue::MergedWithObject,
ConfigValue::MergedWithNonObject,
ConfigValue::ConstructDelayedMerge,
ConfigValue::withFallback

A few of the dependencies of these methods have been stubbed to the
bare minimum for now.